+++ example/lib/add_contact_page.dart
@@ -1,0 +1,107 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
+import 'package:flutter_contact/contacts.dart';
+import 'package:flutter_contact/flutter_contact.dart';
+
+class AddContactPage extends StatefulWidget {
+  @override
+  State<StatefulWidget> createState() => _AddContactPageState();
+}
+
+class _AddContactPageState extends State<AddContactPage> {
+  Contact contact = Contact();
+  PostalAddress address = PostalAddress(label: 'Home');
+  final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Add a contact'),
+        actions: <Widget>[
+          FlatButton(
+            onPressed: () {
+              _formKey.currentState.save();
+              contact.postalAddresses = [address];
+              Contacts.addContact(contact);
+              Navigator.of(context).pop();
+            },
+            child: Icon(Icons.save, color: Colors.white),
+          )
+        ],
+      ),
+      body: Container(
+        padding: EdgeInsets.all(12.0),
+        child: Form(
+          key: _formKey,
+          child: ListView(
+            children: <Widget>[
+              TextFormField(
+                decoration: const InputDecoration(labelText: 'First name'),
+                onSaved: (v) => contact.givenName = v,
+              ),
+              TextFormField(
+                decoration: const InputDecoration(labelText: 'Middle name'),
+                onSaved: (v) => contact.middleName = v,
+              ),
+              TextFormField(
+                decoration: const InputDecoration(labelText: 'Last name'),
+                onSaved: (v) => contact.familyName = v,
+              ),
+              TextFormField(
+                decoration: const InputDecoration(labelText: 'Prefix'),
+                onSaved: (v) => contact.prefix = v,
+              ),
+              TextFormField(
+                decoration: const InputDecoration(labelText: 'Suffix'),
+                onSaved: (v) => contact.suffix = v,
+              ),
+              TextFormField(
+                decoration: const InputDecoration(labelText: 'Phone'),
+                onSaved: (v) => contact.phones = [
+                  if (v != null && v.isNotEmpty) Item(label: 'mobile', value: v)
+                ],
+                keyboardType: TextInputType.phone,
+              ),
+              TextFormField(
+                decoration: const InputDecoration(labelText: 'E-mail'),
+                onSaved: (v) => contact.emails = [
+                  if (v != null && v.isNotEmpty) Item(label: 'work', value: v)
+                ],
+                keyboardType: TextInputType.emailAddress,
+              ),
+              TextFormField(
+                decoration: const InputDecoration(labelText: 'Company'),
+                onSaved: (v) => contact.company = v,
+              ),
+              TextFormField(
+                decoration: const InputDecoration(labelText: 'Job'),
+                onSaved: (v) => contact.jobTitle = v,
+              ),
+              TextFormField(
+                decoration: const InputDecoration(labelText: 'Street'),
+                onSaved: (v) => address.street = v,
+              ),
+              TextFormField(
+                decoration: const InputDecoration(labelText: 'City'),
+                onSaved: (v) => address.city = v,
+              ),
+              TextFormField(
+                decoration: const InputDecoration(labelText: 'Region'),
+                onSaved: (v) => address.region = v,
+              ),
+              TextFormField(
+                decoration: const InputDecoration(labelText: 'Postal code'),
+                onSaved: (v) => address.postcode = v,
+              ),
+              TextFormField(
+                decoration: const InputDecoration(labelText: 'Country'),
+                onSaved: (v) => address.country = v,
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}