+++ example/lib/tiles.dart
@@ -1,0 +1,80 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_contact/contacts.dart';
+
+class ItemsTile extends StatelessWidget {
+  ItemsTile(this._title, this._items);
+
+  final Iterable<Item> _items;
+  final String _title;
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: <Widget>[
+        ListTile(title: Text(_title)),
+        Column(
+          children: _items
+              .map(
+                (i) => Padding(
+                  padding: const EdgeInsets.symmetric(horizontal: 16.0),
+                  child: ListTile(
+                    title: Text(i.label ?? ''),
+                    trailing: Text(i.value ?? ''),
+                  ),
+                ),
+              )
+              .toList(),
+        ),
+      ],
+    );
+  }
+}
+
+class AddressesTile extends StatelessWidget {
+  AddressesTile(this._addresses);
+
+  final Iterable<PostalAddress> _addresses;
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: <Widget>[
+        ListTile(title: Text('Addresses')),
+        Column(
+          children: _addresses
+              .map((a) => Padding(
+                    padding: const EdgeInsets.symmetric(horizontal: 16.0),
+                    child: Column(
+                      children: <Widget>[
+                        ListTile(
+                          title: Text('Street'),
+                          trailing: Text(a.street ?? ''),
+                        ),
+                        ListTile(
+                          title: Text('Postcode'),
+                          trailing: Text(a.postcode ?? ''),
+                        ),
+                        ListTile(
+                          title: Text('City'),
+                          trailing: Text(a.city ?? ''),
+                        ),
+                        ListTile(
+                          title: Text('Region'),
+                          trailing: Text(a.region ?? ''),
+                        ),
+                        ListTile(
+                          title: Text('Country'),
+                          trailing: Text(a.country ?? ''),
+                        ),
+                      ],
+                    ),
+                  ))
+              .toList(),
+        ),
+      ],
+    );
+  }
+}