+++ example/lib/update_person_page.dart
@@ -1,0 +1,149 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:flutter_contact/contacts.dart';
+import 'package:sunny_dart/sunny_dart.dart';
+
+class UpdatePersonPage extends StatefulWidget {
+  UpdatePersonPage({@required this.contact});
+
+  final Contact contact;
+
+  @override
+  _UpdatePersonPageState createState() => _UpdatePersonPageState();
+}
+
+class _UpdatePersonPageState extends State<UpdatePersonPage> {
+  Contact contact;
+  PostalAddress address;
+  Item email;
+  Item phone;
+  final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
+
+  @override
+  void initState() {
+    super.initState();
+    contact = widget.contact;
+    address = contact.postalAddresses.firstOrNull();
+    if (address == null) {
+      address = PostalAddress(label: 'home');
+      contact.postalAddresses.add(address);
+    }
+
+    email = contact.emails.firstOrNull();
+    if (email == null) {
+      email = Item(label: 'home');
+      contact.emails.add(email);
+    }
+
+    phone = contact.phones.firstOrNull();
+    if (phone == null) {
+      phone = Item(label: 'home');
+      contact.phones.add(phone);
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Update Contact'),
+        actions: <Widget>[
+          IconButton(
+            icon: Icon(
+              Icons.save,
+              color: Colors.white,
+            ),
+            onPressed: () async {
+              _formKey.currentState.save();
+              await Contacts.updateContact(contact);
+              await Navigator.of(context).pushReplacementNamed('/');
+            },
+          ),
+        ],
+      ),
+      body: Container(
+        padding: EdgeInsets.all(12.0),
+        child: Form(
+          key: _formKey,
+          child: ListView(
+            children: <Widget>[
+              TextFormField(
+                initialValue: contact.givenName ?? '',
+                decoration: const InputDecoration(labelText: 'First name'),
+                onSaved: (v) => contact.givenName = v,
+              ),
+              TextFormField(
+                initialValue: contact.middleName ?? '',
+                decoration: const InputDecoration(labelText: 'Middle name'),
+                onSaved: (v) => contact.middleName = v,
+              ),
+              TextFormField(
+                initialValue: contact.familyName ?? '',
+                decoration: const InputDecoration(labelText: 'Last name'),
+                onSaved: (v) => contact.familyName = v,
+              ),
+              TextFormField(
+                initialValue: contact.prefix ?? '',
+                decoration: const InputDecoration(labelText: 'Prefix'),
+                onSaved: (v) => contact.prefix = v,
+              ),
+              TextFormField(
+                initialValue: contact.suffix ?? '',
+                decoration: const InputDecoration(labelText: 'Suffix'),
+                onSaved: (v) => contact.suffix = v,
+              ),
+              TextFormField(
+                decoration: const InputDecoration(labelText: 'Phone'),
+                initialValue: phone.value,
+                onSaved: (v) => phone.value = v,
+                keyboardType: TextInputType.phone,
+              ),
+              TextFormField(
+                decoration: const InputDecoration(labelText: 'E-mail'),
+                initialValue: email.value,
+                onSaved: (v) => email.value = v,
+                keyboardType: TextInputType.emailAddress,
+              ),
+              TextFormField(
+                initialValue: contact.company ?? '',
+                decoration: const InputDecoration(labelText: 'Company'),
+                onSaved: (v) => contact.company = v,
+              ),
+              TextFormField(
+                initialValue: contact.jobTitle ?? '',
+                decoration: const InputDecoration(labelText: 'Job'),
+                onSaved: (v) => contact.jobTitle = v,
+              ),
+              TextFormField(
+                initialValue: address.street ?? '',
+                decoration: const InputDecoration(labelText: 'Street'),
+                onSaved: (v) => address.street = v,
+              ),
+              TextFormField(
+                initialValue: address.city ?? '',
+                decoration: const InputDecoration(labelText: 'City'),
+                onSaved: (v) => address.city = v,
+              ),
+              TextFormField(
+                initialValue: address.region ?? '',
+                decoration: const InputDecoration(labelText: 'Region'),
+                onSaved: (v) => address.region = v,
+              ),
+              TextFormField(
+                initialValue: address.postcode ?? '',
+                decoration: const InputDecoration(labelText: 'Postal code'),
+                onSaved: (v) => address.postcode = v,
+              ),
+              TextFormField(
+                initialValue: address.country ?? '',
+                decoration: const InputDecoration(labelText: 'Country'),
+                onSaved: (v) => address.country = v,
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}